Fix for reports out of sync wtith infinite load
Here's what what was happening...

infinite_scroll pager was set to '3', but on the page the main report block was overridden to show 10 items. When pressing "More" the block would show items #4, 5, 6 -> NOT #11, 12, 13.

Evidently infinite_scroll does NOT take the override into account. Even worse, when trying to change the item offset it modifies the original list - it doesn't appear to be possible to have the page display 10 items, but start at the actial offset (11). So... I removed the ability to override this value, and changed infinite_scroll to show 10 items instead of 3.